+++ src/WinRT.Generator.Core/Helpers/MvidGenerator.cs
@@ -6,14 +6,38 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Security.Cryptography;
+using WindowsRuntime.Generator.Extensions;
 
-namespace WindowsRuntime.InteropGenerator.Helpers;
+namespace WindowsRuntime.Generator.Helpers;
 
 /// <summary>
 /// A generator for MVIDs for .NET modules.
 /// </summary>
 internal static class MvidGenerator
 {
+    /// <summary>
+    /// Generates a deterministic MVID based on two input IIDs.
+    /// </summary>
+    /// <param name="left">The first IID to combine.</param>
+    /// <param name="right">The second IID to combine.</param>
+    /// <returns>The resulting MVID.</returns>
+    public static Guid CreateMvid(Guid left, Guid right)
+    {
+        Span<byte> input = stackalloc byte[32];
+
+        // Write the two IIDs in sequence
+        _ = left.TryWriteBytes(input, bigEndian: true, out _);
+        _ = right.TryWriteBytes(input[16..], bigEndian: true, out _);
+
+        Span<byte> hash = stackalloc byte[SHA1.HashSizeInBytes];
+
+        // Hash the two IIDs together (the order matters)
+        _ = SHA1.HashData(input, hash);
+
+        // Create the final MVID from the first 16 bytes of the hash
+        return new(hash[..16]);
+    }
+
     /// <summary>
     /// Generates a deterministic MVID based on a set of input assemblies.
     /// </summary>
